site_master/mysite/requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aioschedule==0.5.2
aiosignal==1.2.0
altgraph==0.17.3
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
auto-py-to-exe==2.23.1
Babel==2.9.1
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
distlib==0.3.5
Django==4.0.6
django-debug-toolbar==3.6.0
djangorestframework==3.14.0
Eel==0.14.0
filelock==3.7.1
fonttools==4.38.0
frozenlist==1.3.1
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.3
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.2
numpy==1.23.3
packaging==21.3
pefile==2022.5.30
Pillow==9.3.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
pywin32-ctypes==0.2.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.15.1
whichcraft==0.6.1
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.4.0
